+++ CHANGE_LOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+ - New Error Event to notify CX regarding Error in widget
  - Padding property to control the padding size choice input adaptive card form field
 
 ## [1.7.3] 2024-11-20